backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
beautifulsoup4
lxml
sentence-transformers
faiss-cpu
numpy
